fix(activity): 把季号标签纯函数挪到客户端安全模块,避免 pg 进客户端 bundle(prod build 修复)+ 文案
上个提交让 "use client" 的 activity-feed.tsx 从 activity-view.ts **值导入** seasonLabelText。
activity-view.ts 有服务端运行时导入(@media-track/workflow → postgres → pg),于是 pg 被打进
浏览器 bundle,Turbopack 无法解析 Node 的 dns/util/types → npm run build:web 报 7 个 module-not-found。
tsc + vitest 都查不出这种 client/server 边界泄漏,只有完整 Next prod build 能抓到。

- 新增 apps/web/lib/activity-season-label.ts:纯函数 distinctSeasons / seasonLabelText 的家,
  只有 import type(EpisodeState / MediaType),零运行时导入 → 可安全被客户端组件值导入。
- activity-feed.tsx:seasonLabelText 改从 ./activity-season-label 导入(ActivityActiveRun 等仍是
  type-only 从 activity-view 导入,客户端安全)。
- activity-view.ts:从纯模块导入并 re-export 这两个函数(服务端调用方与既有导入不破),不再本地定义。
- 单测拆到 activity-season-label.test.ts(从纯模块导入,顺带证明它无运行时依赖);getActivityView
  的多季测试留在 activity-view.test.ts。
- 修 Copilot 文案 nit:seasonNumber 是 snapshot 的占位/锚定季号(全季 run 也可能是真实数字,非 null),
  seasonNumbers 才是实际覆盖季号、即 UI 展示用;不再误称 seasonNumber 对全季为 null。

验证:npm run build:web ✓ Compiled successfully;tsc 绿;vitest 941 passed;eslint 改动文件 0 errors。
